Fix all-static build of dieharder tool.
The all-static dieharder build should be possible by uncommenting
AM_LDFLAGS = -all-static in dieharder/Makefile.am.

Unfortunately it does not work, as order of statically linked
libraries is not correct.

The -lgsl -lgslcblas is already added to LIBS in configure.
So the only change needed is to detect the -lm option that is missing here.

The list dieharder_LDADD is then redundant and can be removed completely.